### 2025-04-29 #7634 @animikhaich
Hi @mahfuz05062, thanks for the invite. However, I am already reviewing another JOSS paper along with a couple other conference papers. Hence, would not have the bandwidth to take on another one at the moment.
