### 2025-05-24 #7634 @mahfuz05062
> Dear Mahfuzur,
>
> When is it due? I could do it but I need about a month. I hope it's
> acceptable
>
> All the best,
>  - Oleg Korobkin
> ___
> Los Alamos National Lab
> Applied Mathematics and Plasma Physics (T-5)
> Phone: (505) 695-8489
> Email: ***@***.***
> […](#)

Hi Oleg, yes, that should be good.
