+++ sbgm/__init__.py
@@ -1,4 +1,3 @@
-from . import _sbgm as sbgm
 from . import sde as sde
 from . import _ode as ode
 from . import _train as train 
