+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "sbgm"
-version = "0.0.22"
+version = "0.0.23"
 description = "Score-based Diffusion models in JAX."
 readme = "README.md"
 requires-python ="~=3.12"
