### 2025-05-30 f279557
exact_log_prob keyword, rademacher noise, examples install packages, version
