+++ configs/mnist.py
@@ -34,7 +34,7 @@ def mnist_config():
 
     # Sampling
     config.sample_size           = 8
-    config.exact_logp            = False
+    config.exact_log_prob        = False
     config.ode_sample            = True
     config.eu_sample             = True
     config.use_ema               = False
