+++ configs/grfs.py
@@ -34,7 +34,7 @@ def grfs_config():
     # Sampling
     config.use_ema         = False
     config.sample_size     = 5
-    config.exact_logp      = False
+    config.exact_log_prob      = False
     config.ode_sample      = True
     config.eu_sample       = True
 
