+++ sbgm/_ode.py
@@ -1,13 +1,23 @@
-from typing import Tuple, Callable, Optional, Sequence, Union
+from typing import Tuple, Callable, Optional, Sequence, Union, Literal
 import jax
 import jax.numpy as jnp
 import jax.random as jr
 import equinox as eqx
 import diffrax as dfx
-from jaxtyping import Key, Array, Float, jaxtyped
+from jaxtyping import PRNGKeyArray, Array, Float, Scalar, jaxtyped
 from beartype import beartype as typechecker
 
 from .sde._sde import SDE
+
+"""
+    Methods for sampling from and computing the log-likelihood of data
+    using a score-based model and a ordinary differential equation (SDE)
+    associated with a stochastic differential equation (SDE).
+    - Hutchinson's trace estimator is used to approximate the log-likelihood
+      if computing the trace directly is not computationally feasible.
+    - The ODE is solved using a numerical solver (e.g., Tsit5) to compute the log-likelihood
+      given a sample x and conditioning variables q and a.
+"""
 
 
 def get_solver() -> dfx.AbstractSolver:
@@ -15,7 +25,7 @@ def get_solver() -> dfx.AbstractSolver:
 
 
 def log_prob_approx(
-    t: Union[float, Float[Array, ""]], 
+    t: Union[float, Scalar], 
     y: Float[Array, "..."], 
     args: Tuple[
         Float[Array, "..."], 
@@ -24,7 +34,7 @@ def log_prob_approx(
         Callable[
             [
                 Float[Array, "..."], 
-                Float[Array, ""], 
+                Scalar, 
                 Optional[Float[Array, "..."]], 
                 Optional[Float[Array, "..."]]
             ], 
@@ -32,7 +42,7 @@ def log_prob_approx(
         ],
         Sequence[int]
     ]
-) -> Tuple[Float[Array, "..."], Float[Array, ""]]:
+) -> Tuple[Float[Array, "..."], Scalar]:
     """ 
         Approx. trace using Hutchinson's trace estimator. 
         - optional multiple-eps sample to average estimated log_prob over
@@ -59,7 +69,7 @@ def log_prob_approx(
 
 
 def log_prob_exact(
-    t: Union[float, Float[Array, ""]], 
+    t: Union[float, Scalar], 
     y: Float[Array, "..."], 
     args: Tuple[
         None, 
@@ -68,7 +78,7 @@ def log_prob_exact(
         Callable[
             [
                 Float[Array, "..."], 
-                Float[Array, ""], 
+                Scalar, 
                 Optional[Float[Array, "..."]], 
                 Optional[Float[Array, "..."]]
             ], 
@@ -76,7 +86,7 @@ def log_prob_exact(
         ],
         Sequence[int]
     ]
-) -> Tuple[Float[Array, "..."], Float[Array, ""]]:
+) -> Tuple[Float[Array, "..."], Scalar]:
     """ 
         Compute trace directly. 
     """
@@ -98,7 +108,7 @@ def get_ode(
 ) -> Callable[
     [
         Float[Array, "..."], 
-        Float[Array, ""], 
+        Scalar, 
         Optional[Float[Array, "..."]], 
         Optional[Float[Array, "..."]]
     ], 
@@ -108,7 +118,7 @@ def get_ode(
 
     def ode(
         y: Float[Array, "..."], 
-        t: Union[float, Float[Array, ""]], 
+        t: Union[float, Scalar], 
         q: Optional[Array] = None, 
         a: Optional[Array] = None
     ) -> Float[Array, "..."]:
@@ -121,17 +131,19 @@ def get_ode(
 @jaxtyped(typechecker=typechecker)
[…]
