+++ configs/flowers.py
@@ -35,7 +35,7 @@ def flowers_config():
 
     # Sampling
     config.sample_size           = 5
-    config.exact_logp            = False
+    config.exact_log_prob            = False
     config.ode_sample            = True
     config.eu_sample             = True
 
