### 2025-06-04 #8347 @editorialbot
[REVIEW]: SBGM: Score-based diffusion models in JAX
**Submitting author:** @homerjed (<a href="http://orcid.org/0009-0002-0985-1437">Jed Homer</a>)
**Repository:** https://github.com/homerjed/sbgm
**Branch with paper.md** (empty if default branch):
**Version:** 0.0.18
