### 2025-07-23 #8347 @korobkin
Hello, my apologies for the late reply. It's been a very busy month. I will
do my best to submit my report by the end of the week

On Sun, Jul 20, 2025, 17:51 Mahfuzur Rahman ***@***.***>
wrote:

> *mahfuz05062* left a comment (openjournals/joss-reviews#8347)
> <https://github.com/openjournals/joss-reviews/issues/8347#issuecomment-3094871033>
>
> Hi @korobkin <https://github.com/korobkin> , I see you haven't started
> your review yet. It's been more than a month now. Can you please reply with
> anything we can help you with to facilitate the process. Thanks!
>
