### 2025-11-09 #8347 @chaitanyaafle
Hi @homerjed, could you please address these minor issues:
1. Include the [year] and the [fullname] in [sbgm/LICENSE](https://github.com/homerjed/sbgm/blob/main/LICENSE)
2. Could you include some guidelines (maybe a line or two in the README.md) for anyone who wishes to contribute to sbgm.
3. In the paper draft, fix one of the references on line 98 (in the section "Implementations and future work").
