### 2025-11-09 #8347 @homerjed
Hi @chaitanyaafle

1. Added both
2. CONTRIBUTING.md plus relevant lines in README.md
3. fixed reference

Thanks! 🙏
