### 2025-11-23 #8347 @korobkin
The package examples seem to require Python 3.12: when I tried installing with the default (latest) version, 3.14, I ran into a problem. Maybe worth mentioning in the README. Lower version doesn't work.

The error looks like this:
```
$ pip install .[examples]
. . . < skipped > . . .
NFO: pip is looking at multiple versions of sbgm to determine which version is compatible with other requirements. This could take a while.
ERROR: Ignored the following versions that require a different python version: 1.21.2 Requires-Python >=3.7,<3.11; 1.21.3 Requires-Python >=3.7,<3.11; 1.21.4 Requires-Python >=3.7,<3.11; 1.21.5 Requires-Python >=3.7,<3.11; 1.21.6 Requires-Python >=3.7,<3.11; 1.26.0 Requires-Python >=3.9,<3.13; 1.26.1 Requires-Python >=3.9,<3.13
ERROR: Could not find a version that satisfies the requirement torch==2.4.0 (from sbgm) (from versions: 2.9.0, 2.9.1)
ERROR: No matching distribution found for torch==2.4.0
```

With python 3.12, everything installs fine.
