### 2025-11-24 #8347 @homerjed
Thank you @korobkin for your review! Much appreciated 🙏.

My initial response here is to leave things as they are especially as now the checklists are complete.

For python 3.13, 3.14 the corresponding `jax` versions are dropping support for `cuda12` accelerators. This is a problem for most academic labs who aren't purchasing the newest GPUs etc.

Once I'm done with my PhD my plan would be to migrate completely to the hugging face `datasets` API (to drop `torch`). This would be agnostic to your `torch`/`jax`-accelerator setup. Additionally, I would move to `uv` package management to ensure this works on >3.12 (this is needlessly difficult to implement without `uv` imo).
