### 2025-12-29 #8347 @homerjed
Hi @crvernon,

Thank you!

Archive
- Changed archive title to exactly the same as the paper.
- Changed license in Zenodo archive to MIT, license in github repo is MIT.

Paper
- Line 139 and Line 165 'bayesian' -> 'Bayesian'
- Yep that was an error, fixed to the single bib entry for that paper.

I've checked these are all corrected, all good from my side I think. Thanks again.
